re #63 expose session id as uLong

+++ src/NEventSocket/Channels/ChannelVariables.cs
@@ -17,8 +17,15 @@ namespace NEventSocket.Channels
 
         public string this[string variableName] => this.channel.GetVariable(variableName);
 
+        /// <summary>
+        /// Gets hangup reason for the last bridge attempt
+        /// </summary>
+        /// <remarks>
+        /// last_bridge_hangup_cause is not populated in certain cases eg. USER_NOT_REGISTERED
+        /// will check originate_disposition if not present
+        /// </remarks>
         public HangupCause? BridgeHangupCause => (this["last_bridge_hangup_cause"] ?? this["originate_disposition"]).HeaderToEnumOrNull<HangupCause>();
 
-        public int SessionId => int.Parse(this["session_id"]);
+        public ulong SessionId => ulong.Parse(this["session_id"]);
     }
 }
